Addition of language changing
This updated includes a new language handler that will store and update the text of the main index page. The module uses dictionaries to store various languages. The new functionality also required some modification to the index, css, and the crossword-puzzel script. Major changes included the addition of the i18n tag for language translation, and the change to the crossword generator to allow for different languages.